Make sure order of analyses is saved properly. (#3986)
- Fixes https://github.com/jasp-stats/jasp-issues/issues/639
- Also moved all the functions for moving analyses together and into Analyses, Im not sure why part of it was on MainWindow.

+++ JASP-Desktop/results/resultsjsinterface.cpp
@@ -270,7 +270,7 @@ void ResultsJsInterface::removeAnalyses()
 	emit runJavaScript("window.removeAllAnalyses()");
 }
 
-void ResultsJsInterface::moveAnalyses(size_t fromId, size_t toId)
+void ResultsJsInterface::moveAnalyses(quint64 fromId, quint64 toId)
 {
 	emit runJavaScript("window.moveAnalyses(" % QString::number(fromId) % "," % QString::number(toId) % ")");
 }
